fix pr issues
