+++ Pokedex/src/pages/Home/HomePage.tsx
@@ -7,20 +7,22 @@ interface HomePageProps {
   setActivePage: (page: string) => void,
 }
 
+const getInitialRecentSearches = () => {
+  const savedSearches = localStorage.getItem('recentSearches');
+  return savedSearches ? JSON.parse(savedSearches) : [];
+};
+
 function HomePage({ setActivePage } : HomePageProps) {
   setActivePage(HOME_PAGE_TITLE);
+  
+  const [recentSearches, setRecentSearches] = useState<string[]>(getInitialRecentSearches);
 
-  const [recentSearches, setRecentSearches] = useState<string[]>(() => {
-    const savedSearches = localStorage.getItem('recentSearches');
-    return savedSearches ? JSON.parse(savedSearches) : [];
-  });
-
-  const [selectedOption, setSelectedOption] = useState('');
+  const [selectedPokemon, setSelectedPokemon] = useState('');
 
   return (
     <>
-      <DropDown options={recentSearches} selectedOption={selectedOption} setSelectedOption={setSelectedOption} />
-      <PokeGrid selectedOption={selectedOption} />
+      <DropDown options={recentSearches} selectedOption={selectedPokemon} setSelectedOption={setSelectedPokemon} />
+      <PokeGrid selectedOption={selectedPokemon} />
     </>
   )
 }
